+++ NEWS.md
@@ -1,4 +1,6 @@
-# tectonicr 0.2.98
+# tectonicr 0.2.98 _2024-04-07_
+
+* minor fixes
 
 # tectonicr 0.2.97 _2024-04-07_